Fixed some bugs in relation to modularity; - Removed runtime resource validation and put it under tests instead; Build.gradle: - Added forceMerge in jlink task, since log4j is not designed for modularity yet; - Renamed the group name, to not have trailing numbers.
