+++ build.gradle
@@ -4,7 +4,7 @@ plugins {
     id "org.beryx.jlink" version "2.24.4"
 }
 
-group = 'nl.danman85'
+group = 'nl.danman'
 version = '0.0.1'
 
 sourceCompatibility = 11
@@ -39,6 +39,11 @@ javafx {
             'javafx.controls',
             'javafx.fxml',
     ]
+}
+
+jlink {
+    options = ['--strip-debug', '--compress', '2', '--no-header-files', '--no-man-pages']
+    forceMerge 'log4j-api'
 }
 
 tasks.named('test') {
