+++ src/main/java/nl/danman/file_encryptor/client/views/FXMLViewDefinition.java
@@ -1,12 +1,11 @@
-package nl.danman.file_encryptor.client.configuration;
+package nl.danman.file_encryptor.client.views;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
-import nl.danman.file_encryptor.client.views.Controller;
 import nl.danman.file_encryptor.client.views.file_encryptor.FileEncryptorController;
 import nl.danman.file_encryptor.client.views.file_encryptor.FileLineController;
 import nl.danman.file_encryptor.client.views.main.MainViewController;
 
-public enum FXMLViewSource {
+public enum FXMLViewDefinition {
 
     MAIN_VIEW("/views/MainView.fxml", MainViewController.class),
     FILE_ENCRYPTOR("/views/file_encryptor/FileEncryptor.fxml", FileEncryptorController.class),
@@ -16,7 +15,7 @@ public enum FXMLViewSource {
 
     private final Class<? extends Controller> clazz;
 
-    FXMLViewSource(final String url, final Class<? extends Controller> clazz) {
+    FXMLViewDefinition(final String url, final Class<? extends Controller> clazz) {
         this.resourceUrl = url;
         this.clazz = clazz;
     }
